Add struct type for ZMTP commands.
The struct added is for the NULL authentification protocol.
It can be parsed from a stream of bytes using `FrameType::with_stream`.
It returns a `RawFrame` on which we implemented `From<RawFrae> for Frame`.
NODE: this indirection is probably useless, and puting the full Frame in a buffer before the parse may be an issue with big Frames.
Later, we will probably try to have a stream oriented API wherever it's possible.