rename `Frame::Empty` to `Frame::Separator` to be aligned with the specifications.
In the specs, it's said, a separator is an empty frame but without defining exactly the type of frame it is suposed to be.
In practice, what is used as a separator is a short message with the byte `MORE` up and a size of 0.
It's not suposed to be the only type of empty frame so we prefer to name it a separator.